+++ src/libtcod/parser_c.c
@@ -37,18 +37,13 @@
 #include "libtcod_int.h"
 #include "parser.h"
 
-#define BIG_NAME_LEN 128
-
-#define DEF_NAME(d) (((TCOD_struct_int_t*)d)->name)
-#define DEF_FLAGS(d) (((TCOD_struct_int_t*)d)->flags)
-#define DEF_PROPS(d) (((TCOD_struct_int_t*)d)->props)
-#define DEF_LISTS(d) (((TCOD_struct_int_t*)d)->lists)
+#define MAX_NAME_LEN 128
 
 /* definition property */
-typedef struct {
+typedef struct TCOD_struct_prop_t {
   char* name;
   TCOD_value_type_t value;
-  bool mandat;
+  bool mandatory;
 } TCOD_struct_prop_t;
 
 static TCOD_lex_t* lex = NULL;
@@ -59,16 +54,16 @@ static const char* keywords[] = {"struct", "bool", "char", "int", "float", "stri
 
 static TCOD_parser_listener_t* listener = NULL;
 
-static bool default_new_struct(TCOD_parser_struct_t str, const char* name);
+static bool default_new_struct(TCOD_ParserStruct* str, const char* name);
 static bool default_new_flag(const char* name);
 static bool default_new_property(const char* propname, TCOD_value_type_t type, TCOD_value_t value);
-static bool default_end_struct(TCOD_parser_struct_t str, const char* name);
+static bool default_end_struct(TCOD_ParserStruct* str, const char* name);
 static void default_error(const char* msg);
 
 static TCOD_parser_listener_t default_listener = {
     default_new_struct, default_new_flag, default_new_property, default_end_struct, default_error};
 
-static bool string_copy(char* dest, char* source, int len) {
+static bool string_copy(char* dest, const char* source, int len) {
   if (source == NULL) return false;
   strncpy(dest, source, len);
   dest[len - 1] = '\0';
@@ -88,44 +83,44 @@ void TCOD_parser_error(const char* msg, ...) {
   lex->token_type = TCOD_LEX_ERROR;
 }
 
-const char* TCOD_struct_get_name(TCOD_parser_struct_t def) { return ((TCOD_struct_int_t*)def)->name; }
+const char* TCOD_struct_get_name(const TCOD_ParserStruct* def) { return def->name; }
 
 /* add a property to an entity definition */
-void TCOD_struct_add_property(TCOD_parser_struct_t def, const char* name, TCOD_value_type_t type, bool mandatory) {
-  TCOD_struct_prop_t* prop = (TCOD_struct_prop_t*)calloc(1, sizeof(TCOD_struct_prop_t));
+void TCOD_struct_add_property(TCOD_ParserStruct* def, const char* name, TCOD_value_type_t type, bool mandatory) {
+  TCOD_struct_prop_t* prop = calloc(1, sizeof(*prop));
   prop->name = TCOD_strdup(name);
   prop->value = type;
-  prop->mandat = mandatory;
-  TCOD_list_push(DEF_PROPS(def), (void*)prop);
+  prop->mandatory = mandatory;
+  TCOD_list_push(def->props, prop);
 }
 
 /* add a list property to an entity definition */
-void TCOD_struct_add_list_property(TCOD_parser_struct_t def, const char* name, TCOD_value_type_t type, bool mandatory) {
+void TCOD_struct_add_list_property(TCOD_ParserStruct* def, const char* name, TCOD_value_type_t type, bool mandatory) {
   TCOD_struct_add_property(def, name, type | TCOD_TYPE_LIST, mandatory);
 }
 
 void TCOD_struct_add_value_list_sized(
-    TCOD_parser_struct_t def, const char* name, const char** value_list, int size, bool mandatory) {
-  char** newArray = NULL;
+    TCOD_ParserStruct* def, const char* name, const char* const* value_list, int size, bool mandatory) {
   /* duplicate value list to avoid issues with C# garbage collector */
-  TCOD_value_type_t type = (TCOD_value_type_t)((int)TCOD_TYPE_VALUELIST00 + TCOD_list_size(DEF_LISTS(def)));
-  int i = 0;
+  const int type = TCOD_TYPE_VALUELIST00 + TCOD_list_size(def->lists);
 
-  if (size) newArray = calloc(size + 1, sizeof(char*));
+  char** newArray = NULL;
+  if (size) newArray = calloc(size + 1, sizeof(*newArray));
 
-  for (i = 0; i < size; i++) newArray[i] = TCOD_strdup(value_list[i]);
+  for (int i = 0; i < size; ++i) newArray[i] = TCOD_strdup(value_list[i]);
   newArray[size] = NULL;
 
   TCOD_struct_add_property(def, name, type, mandatory);
-  TCOD_list_push(DEF_LISTS(def), (void*)newArray);
+  TCOD_list_push(def->lists, newArray);
 }
 
 /* add a value-list property to an entity definition */
-void TCOD_struct_add_value_list(TCOD_parser_struct_t def, const char* name, const char** value_list, bool mandatory) {
+void TCOD_struct_add_value_list(
+    TCOD_ParserStruct* def, const char* name, const char* const* value_list, bool mandatory) {
   int size = 0;
   if (value_list) {
     while (value_list[size] != NULL) {
-      size++;
+      ++size;
     };
   }
   TCOD_struct_add_value_list_sized(def, name, value_list, size, mandatory);
@@ -133,106 +128,85 @@ void TCOD_struct_add_value_list(TCOD_parser_struct_t def, const char* name, cons
 
 /* add a flag (simplified bool value) to an entity definition */
 /* a flag cannot be mandatory. if present => true, if omitted => false */
-void TCOD_struct_add_flag(TCOD_parser_struct_t def, const char* propname) {
-  TCOD_list_push(DEF_FLAGS(def), (void*)TCOD_strdup(propname));
+void TCOD_struct_add_flag(TCOD_ParserStruct* def, const char* propname) {
+  TCOD_list_push(def->flags, TCOD_strdup(propname));
 }
 
 /* add a sub-entity to an entity definition */
-void TCOD_struct_add_structure(TCOD_parser_struct_t def, TCOD_parser_struct_t sub_definition) {
-  TCOD_list_push(((TCOD_struct_int_t*)def)->structs, (const void*)sub_definition);
+void TCOD_struct_add_structure(TCOD_ParserStruct* def, const TCOD_ParserStruct* sub_definition) {
+  TCOD_list_push(def->structs, sub_definition);
 }
 
 /* check if given property is mandatory for this entity definition */
-bool TCOD_struct_is_mandatory(TCOD_parser_struct_t def, const char* propname) {
-  TCOD_struct_prop_t** i_prop;
-  for (i_prop = (TCOD_struct_prop_t**)TCOD_list_begin(DEF_PROPS(def));
-       i_prop != (TCOD_struct_prop_t**)TCOD_list_end(DEF_PROPS(def));
+bool TCOD_struct_is_mandatory(TCOD_ParserStruct* def, const char* propname) {
+  for (TCOD_struct_prop_t** i_prop = (void*)TCOD_list_begin(def->props); i_prop != (void*)TCOD_list_end(def->props);
        i_prop++) {
-    if (strcmp((*i_prop)->name, propname) == 0) return (*i_prop)->mandat;
+    if (strcmp((*i_prop)->name, propname) == 0) return (*i_prop)->mandatory;
   }
   return false;
 }
 
 /* returns the type of given property */
 /* NONE if the property does not exist */
-TCOD_value_type_t TCOD_struct_get_type(TCOD_parser_struct_t def, const char* propname) {
-  TCOD_struct_prop_t** i_prop;
-  char** iflag;
-  for (i_prop = (TCOD_struct_prop_t**)TCOD_list_begin(DEF_PROPS(def));
-       i_prop != (TCOD_struct_prop_t**)TCOD_list_end(DEF_PROPS(def));
+TCOD_value_type_t TCOD_struct_get_type(const TCOD_ParserStruct* def, const char* propname) {
+  for (const TCOD_struct_prop_t** i_prop = (void*)TCOD_list_begin(def->props);
+       i_prop != (void*)TCOD_list_end(def->props);
        i_prop++) {
     if (strcmp((*i_prop)->name, propname) == 0) return (*i_prop)->value;
   }
-  for (iflag = (char**)TCOD_list_begin(DEF_FLAGS(def)); iflag != (char**)TCOD_list_end(DEF_FLAGS(def)); iflag++) {
+  for (const char** iflag = (void*)TCOD_list_begin(def->flags); iflag != (void*)TCOD_list_end(def->flags); iflag++) {
     if (strcmp((*iflag), propname) == 0) return TCOD_TYPE_BOOL;
   }
   return TCOD_TYPE_NONE;
 }
 
 TCOD_value_t TCOD_parse_bool_value(void) {
-  TCOD_value_t ret;
-  ret.b = false;
-  if (strcmp(lex->tok, "true") == 0)
-    ret.b = true;
-  else if (strcmp(lex->tok, "false") == 0)
-    ret.b = false;
-  else
-    TCOD_parser_error("parseBoolValue : unknown value %s for bool. 'true' or 'false' expected", lex->tok);
-  return ret;
+  if (strcmp(lex->tok, "true") == 0) return (TCOD_value_t){.b = true};
+  if (strcmp(lex->tok, "false") == 0) return (TCOD_value_t){.b = false};
+  TCOD_parser_error("parseBoolValue : unknown value %s for bool. 'true' or 'false' expected", lex->tok);
+  return (TCOD_value_t){.b = false};
 }
 
 TCOD_value_t TCOD_parse_char_value(void) {
-  TCOD_value_t ret;
   if (lex->token_type != TCOD_LEX_CHAR && lex->token_type != TCOD_LEX_INTEGER) {
     TCOD_parser_error("parseCharValue : char constant expected instead of '%s'", lex->tok);
   }
-  ret.c = (char)lex->token_int_val;
-  return ret;
+  return (TCOD_value_t){.c = (char)lex->token_int_val};
 }
 
 TCOD_value_t TCOD_parse_integer_value(void) {
-  TCOD_value_t ret;
   if (lex->token_type != TCOD_LEX_INTEGER)
     TCOD_parser_error("parseIntegerValue : integer constant expected instead of '%s'", lex->tok);
-  ret.i = lex->token_int_val;
-  return ret;
+  return (TCOD_value_t){.i = lex->token_int_val};
 }
 
 TCOD_value_t TCOD_parse_float_value(void) {
-  TCOD_value_t ret;
   if (lex->token_type != TCOD_LEX_FLOAT && lex->token_type != TCOD_LEX_INTEGER)
     TCOD_parser_error("parseFloatValue : float constant expected instead of '%s'", lex->tok);
-  if (lex->token_type == TCOD_LEX_FLOAT)
-    ret.f = lex->token_float_val;
-  else
-    ret.f = (float)lex->token_int_val;
-  return ret;
+  return (TCOD_value_t){.f = lex->token_type == TCOD_LEX_FLOAT ? lex->token_float_val : (float)lex->token_int_val};
 }
 
 TCOD_value_t TCOD_parse_string_value(void) {
-  TCOD_value_t ret;
-  TCOD_list_t l;
-  bool end = false;
-  char** s;
-  size_t slen = 0;
-  l = TCOD_list_new();
+  TCOD_list_t l = TCOD_list_new();
   if (lex->token_type != TCOD_LEX_STRING)
     TCOD_parser_error("parseStringValue : string constant expected instead of '%s'", lex->tok);
-  while (!end) {
+  while (true) {
     TCOD_lex_t save;
-    TCOD_list_push(l, (void*)TCOD_strdup(lex->tok));
+    TCOD_list_push(l, TCOD_strdup(lex->tok));
     TCOD_lex_savepoint(lex, &save);
     if (TCOD_lex_parse(lex) != TCOD_LEX_STRING) {
-      end = true;
       TCOD_lex_restore(lex, &save);
+      break;
     }
   }
-  for (s = (char**)TCOD_list_begin(l); s != (char**)TCOD_list_end(l); s++) {
+  size_t slen = 0;
+  for (char** s = (void*)TCOD_list_begin(l); s != (void*)TCOD_list_end(l); ++s) {
     slen += strlen(*s);
   }
-  ret.s = (char*)calloc(sizeof(char), slen + 1);
-  for (s = (char**)TCOD_list_begin(l); s != (char**)TCOD_list_end(l); s++) {
-    strcat(ret.s, *s);
+  TCOD_value_t ret = {.s = calloc(sizeof(*ret.s), slen + 1)};
+  if (!ret.s) TCOD_parser_error("parseStringValue : out of memory allocating string of length %d.", slen + 1);
+  for (char** s = (void*)TCOD_list_begin(l); s != (void*)TCOD_list_end(l); ++s) {
+    if (ret.s) strcat(ret.s, *s);
     free(*s);
   }
   TCOD_list_delete(l);
@@ -240,8 +214,7 @@ TCOD_value_t TCOD_parse_string_value(void) {
 }
 
 TCOD_value_t TCOD_parse_color_value(void) {
-  TCOD_value_t ret;
-  ret.col = (TCOD_ColorRGB){0, 0, 0};
+  TCOD_value_t ret = {.col = {0, 0, 0}};
   if (lex->token_type == TCOD_LEX_SYMBOL && lex->tok[0] == '#') {
     char tmp[128] = "";
     int tok = TCOD_lex_parse(lex);
@@ -265,13 +238,12 @@ TCOD_value_t TCOD_parse_color_value(void) {
   if (lex->token_type != TCOD_LEX_STRING)
     TCOD_parser_error("parseColorValue : string constant expected instead of '%s'", lex->tok);
   if (lex->tok[0] == '#') {
-    int r, g, b;
     if (strlen(lex->tok) != 7)
       TCOD_parser_error("parseColorValue : bad color format. '#rrggbb' expected instead of '%s'", lex->tok);
     /* web format : #rrggbb */
-    r = (TCOD_lex_hextoint(lex->tok[1]) << 4) + TCOD_lex_hextoint(lex->tok[2]);
-    g = (TCOD_lex_hextoint(lex->tok[3]) << 4) + TCOD_lex_hextoint(lex->tok[4]);
-    b = (TCOD_lex_hextoint(lex->tok[5]) << 4) + TCOD_lex_hextoint(lex->tok[6]);
+    const int r = (TCOD_lex_hextoint(lex->tok[1]) << 4) + TCOD_lex_hextoint(lex->tok[2]);
+    const int g = (TCOD_lex_hextoint(lex->tok[3]) << 4) + TCOD_lex_hextoint(lex->tok[4]);
+    const int b = (TCOD_lex_hextoint(lex->tok[5]) << 4) + TCOD_lex_hextoint(lex->tok[6]);
     ret.col.r = (uint8_t)r;
     ret.col.g = (uint8_t)g;
     ret.col.b = (uint8_t)b;
@@ -302,14 +274,9 @@ TCOD_value_t TCOD_parse_color_value(void) {
 
 TCOD_value_t TCOD_parse_dice_value(void) {
   /* dice format : [<m>(x|*)]<n>(D|d)<f>[(+|-)<a>] */
-  TCOD_value_t ret;
-  bool minus = false;
-  char* begin;
-  char* ptr;
-  ret.dice.multiplier = 1.0f;
-  ret.dice.addsub = 0.0f;
-  begin = lex->tok;
-  ptr = strchr(begin, 'x');
+  TCOD_value_t ret = {.dice = {.multiplier = 1.0f}};
+  char* begin = lex->tok;
+  char* ptr = strchr(begin, 'x');
   if (!ptr) ptr = strchr(begin, '*');
   if (ptr) {
     /* parse multiplier */
@@ -327,6 +294,7 @@ TCOD_value_t TCOD_parse_dice_value(void) {
   ret.dice.nb_rolls = atoi(begin);
   begin = ptr + 1;
   ptr = strchr(begin, '+');
+  bool minus = false;
   if (!ptr) {
     ptr = strchr(begin, '-');
     if (ptr) minus = true;
@@ -343,13 +311,12 @@ TCOD_value_t TCOD_parse_dice_value(void) {
   return ret;
 }
 
-TCOD_value_t TCOD_parse_value_list_value(TCOD_struct_int_t* def, int listnum) {
-  TCOD_value_t ret;
-  int i = 0;
+TCOD_value_t TCOD_parse_value_list_value(TCOD_ParserStruct* def, int list_num) {
   char* value;
-  char** value_list = TCOD_list_get(def->lists, listnum);
+  char** value_list = TCOD_list_get(def->lists, list_num);
   if (lex->token_type != TCOD_LEX_STRING)
     TCOD_parser_error("parseValueListValue : string constant expected instead of '%s'", lex->tok);
+  int i = 0;
   value = value_list[i];
   while (value) {
     if (strcmp(lex->tok, value) == 0) break;
@@ -357,14 +324,13 @@ TCOD_value_t TCOD_parse_value_list_value(TCOD_struct_int_t* def, int listnum) {
     value = value_list[i];
   }
   if (!value) TCOD_parser_error("parseValueListValue : incorrect value '%s'", lex->tok);
-  ret.s = value;
-  return ret;
+  return (TCOD_value_t){.s = value};
 }
 
-TCOD_value_t TCOD_parse_property_value(TCOD_parser_int_t* parser, TCOD_parser_struct_t def, char* propname, bool list) {
+TCOD_value_t TCOD_parse_property_value(TCOD_Parser* parser, TCOD_ParserStruct* def, char* propname, bool list) {
   TCOD_value_type_t type = TCOD_struct_get_type(def, propname);
-  TCOD_value_t ret = {0};
   if (!list) type &= ~TCOD_TYPE_LIST;
+  TCOD_value_t ret = {0};
   if (type & TCOD_TYPE_LIST) {
     type &= ~TCOD_TYPE_LIST;
     if (strcmp(lex->tok, "[") != 0) {
@@ -372,12 +338,11 @@ TCOD_value_t TCOD_parse_property_value(TCOD_parser_int_t* parser, TCOD_parser_st
     }
     ret.list = TCOD_list_new();
     do {
-      TCOD_value_t val;
-      int tok = TCOD_lex_parse(lex);
+      const int tok = TCOD_lex_parse(lex);
       if (tok == TCOD_LEX_EOF || tok == TCOD_LEX_ERROR) {
         TCOD_parser_error("Missing ']' in list value");
       }
-      val = TCOD_parse_property_value(parser, def, propname, false);
+      TCOD_value_t val = TCOD_parse_property_value(parser, def, propname, false);
       if (type == TCOD_TYPE_STRING || (type >= TCOD_TYPE_VALUELIST00 && type <= TCOD_TYPE_VALUELIST15)) {
         TCOD_list_push(ret.list, TCOD_strdup(val.s));
       } else {
@@ -392,25 +357,18 @@ TCOD_value_t TCOD_parse_property_value(TCOD_parser_int_t* parser, TCOD_parser_st
     switch (type) {
       case TCOD_TYPE_BOOL:
         return TCOD_parse_bool_value();
-        break;
       case TCOD_TYPE_CHAR:
         return TCOD_parse_char_value();
-        break;
       case TCOD_TYPE_INT:
         return TCOD_parse_integer_value();
-        break;
       case TCOD_TYPE_FLOAT:
         return TCOD_parse_float_value();
-        break;
       case TCOD_TYPE_STRING:
         return TCOD_parse_string_value();
-        break;
       case TCOD_TYPE_COLOR:
         return TCOD_parse_color_value();
-        break;
       case TCOD_TYPE_DICE:
         return TCOD_parse_dice_value();
-        break;
       case TCOD_TYPE_VALUELIST00:
       case TCOD_TYPE_VALUELIST01:
       case TCOD_TYPE_VALUELIST02:
@@ -426,11 +384,8 @@ TCOD_value_t TCOD_parse_property_value(TCOD_parser_int_t* parser, TCOD_parser_st
       case TCOD_TYPE_VALUELIST12:
       case TCOD_TYPE_VALUELIST13:
       case TCOD_TYPE_VALUELIST14:
-      case TCOD_TYPE_VALUELIST15: {
-        int listnum = type - TCOD_TYPE_VALUELIST00;
-        return TCOD_parse_value_list_value((TCOD_struct_int_t*)def, listnum);
-        break;
-      }
+      case TCOD_TYPE_VALUELIST15:
+        return TCOD_parse_value_list_value(def, type - TCOD_TYPE_VALUELIST00);
       case TCOD_TYPE_CUSTOM00:
       case TCOD_TYPE_CUSTOM01:
       case TCOD_TYPE_CUSTOM02:
@@ -453,21 +408,21 @@ TCOD_value_t TCOD_parse_property_value(TCOD_parser_int_t* parser, TCOD_parser_st
           TCOD_parser_error(
               "parse_property_value : no custom parser for property type %d for entity %s prop %s",
               type,
-              DEF_NAME(def),
+              def->name,
               propname);
         }
         break;
       default:
         TCOD_parser_error(
-            "parse_property_value : unknown property type %d for entity %s prop %s", type, DEF_NAME(def), propname);
+            "parse_property_value : unknown property type %d for entity %s prop %s", type, def->name, propname);
         break;
     }
   }
   return ret;
 }
 
-static bool TCOD_parser_parse_entity(TCOD_parser_int_t* parser, TCOD_struct_int_t* def) {
-  char* name = NULL;
+static bool TCOD_parser_parse_entity(TCOD_Parser* parser, TCOD_ParserStruct* def) {
+  const char* name = NULL;
   if (TCOD_lex_parse(lex) == TCOD_LEX_STRING) {
     /* entity type name */
     name = TCOD_strdup(lex->tok);
@@ -479,8 +434,6 @@ static bool TCOD_parser_parse_entity(TCOD_parser_int_t* parser, TCOD_struct_int_
   }
   TCOD_lex_parse(lex);
   while (strcmp(lex->tok, "}") != 0) {
-    bool found = false;
-    char** iflag;
     bool dynStruct = false;
     TCOD_value_type_t dynType = TCOD_TYPE_NONE;
     if (lex->token_type == TCOD_LEX_KEYWORD) {
@@ -527,8 +480,9 @@ static bool TCOD_parser_parse_entity(TCOD_parser_int_t* parser, TCOD_struct_int_
       return false;
     }
     /* is it a flag ? */
+    bool found = false;
     if (!dynStruct && dynType == TCOD_TYPE_NONE) {
-      for (iflag = (char**)TCOD_list_begin(def->flags); iflag != (char**)TCOD_list_end(def->flags); iflag++) {
+      for (char** iflag = (void*)TCOD_list_begin(def->flags); iflag != (void*)TCOD_list_end(def->flags); ++iflag) {
         if (strcmp(*iflag, lex->tok) == 0) {
           found = true;
           if (!listener->new_flag(lex->tok)) return false;
@@ -544,8 +498,8 @@ static bool TCOD_parser_parse_entity(TCOD_parser_int_t* parser, TCOD_struct_int_
              i_prop != (TCOD_struct_prop_t**)TCOD_list_end(def->props);
              i_prop++) {
           if (strcmp((*i_prop)->name, lex->tok) == 0) {
-            char propname[BIG_NAME_LEN];
-            string_copy(propname, lex->tok, BIG_NAME_LEN);
+            char propname[MAX_NAME_LEN];
+            string_copy(propname, lex->tok, MAX_NAME_LEN);
             TCOD_lex_parse(lex);
             if (strcmp(lex->tok, "=") != 0) {
               TCOD_parser_error("Parser::parseEntity : '=' expected");
@@ -570,26 +524,25 @@ static bool TCOD_parser_parse_entity(TCOD_parser_int_t* parser, TCOD_struct_int_
     }
     if (!found) {
       /* is it a sub-entity type */
-      char id[BIG_NAME_LEN * 2 + 2];
+      char id[MAX_NAME_LEN * 2 + 2];
       bool blockFound = false;
       do {
         TCOD_lex_t save;
-        char type[BIG_NAME_LEN];
-        char* subname = NULL;
-        bool named = false;
         TCOD_lex_savepoint(lex, &save);
-        string_copy(type, lex->tok, BIG_NAME_LEN);
+        char type[MAX_NAME_LEN];
+        string_copy(type, lex->tok, MAX_NAME_LEN);
         strcpy(id, type);
+        const char* subname = NULL;
+        bool named = false;
         if (TCOD_lex_parse(lex) == TCOD_LEX_STRING) {
           /* <type>#<name> */
-          TCOD_struct_int_t** sub;
           strcat(id, "#");
           strcat(id, lex->tok);
           named = true;
           subname = TCOD_strdup(lex->tok);
           TCOD_lex_restore(lex, &save);
-          for (sub = (TCOD_struct_int_t**)TCOD_list_begin(def->structs);
-               sub != (TCOD_struct_int_t**)TCOD_list_end(def->structs);
+          for (TCOD_ParserStruct** sub = (void*)TCOD_list_begin(def->structs);
+               sub != (void*)TCOD_list_end(def->structs);
                sub++) {
             if (strcmp((*sub)->name, id) == 0) {
               if (!listener->new_struct(*sub, lex->tok)) return false;
@@ -604,9 +557,8 @@ static bool TCOD_parser_parse_entity(TCOD_parser_int_t* parser, TCOD_struct_int_
         }
         if (!blockFound) {
           /* <type> alone */
-          TCOD_struct_int_t** sub;
-          for (sub = (TCOD_struct_int_t**)TCOD_list_begin(def->structs);
-               sub != (TCOD_struct_int_t**)TCOD_list_end(def->structs);
+          for (TCOD_ParserStruct** sub = (void*)TCOD_list_begin(def->structs);
+               sub != (void*)TCOD_list_end(def->structs);
                sub++) {
             if (strcmp((*sub)->name, type) == 0) {
               if (!listener->new_struct(*sub, subname)) return false;
@@ -619,10 +571,9 @@ static bool TCOD_parser_parse_entity(TCOD_parser_int_t* parser, TCOD_struct_int_
         }
         if (!blockFound && dynStruct) {
           /* unknown structure. auto-declaration */
-          TCOD_struct_int_t** i_def;
-          TCOD_struct_int_t* s = NULL;
-          for (i_def = (TCOD_struct_int_t**)TCOD_list_begin(parser->structs);
-               i_def != (TCOD_struct_int_t**)TCOD_list_end(parser->structs);
+          const TCOD_ParserStruct* s = NULL;
+          for (const TCOD_ParserStruct** i_def = (void*)TCOD_list_begin(parser->structs);
+               i_def != (void*)TCOD_list_end(parser->structs);
                i_def++) {
             if (strcmp((*i_def)->name, id) == 0) {
               s = *i_def;
@@ -631,8 +582,8 @@ static bool TCOD_parser_parse_entity(TCOD_parser_int_t* parser, TCOD_struct_int_
           }
           if (s == NULL && named) {
             /* look for general definition <type> for entity <type>#<name> */
-            for (i_def = (TCOD_struct_int_t**)TCOD_list_begin(parser->structs);
-                 i_def != (TCOD_struct_int_t**)TCOD_list_end(parser->structs);
+            for (const TCOD_ParserStruct** i_def = (void*)TCOD_list_begin(parser->structs);
+                 i_def != (void*)TCOD_list_end(parser->structs);
                  i_def++) {
               if (strcmp((*i_def)->name, type) == 0) {
                 s = *i_def;
@@ -662,76 +613,68 @@ static bool TCOD_parser_parse_entity(TCOD_parser_int_t* parser, TCOD_struct_int_
 /* generic parser */
 /****************************************/
 
-TCOD_parser_t TCOD_parser_new(void) {
-  TCOD_parser_int_t* ent = (TCOD_parser_int_t*)calloc(1, sizeof(TCOD_parser_int_t));
+TCOD_Parser* TCOD_parser_new(void) {
+  TCOD_Parser* ent = calloc(1, sizeof(*ent));
   ent->structs = TCOD_list_new();
-  return (TCOD_parser_t)ent;
+  return ent;
 }
 
-TCOD_value_type_t TCOD_parser_new_custom_type(TCOD_parser_t parser, TCOD_parser_custom_t custom_type_parser) {
-  TCOD_parser_int_t* p = (TCOD_parser_int_t*)parser;
+TCOD_value_type_t TCOD_parser_new_custom_type(TCOD_Parser* parser, TCOD_parser_custom_t custom_type_parser) {
   TCOD_value_type_t type = TCOD_TYPE_CUSTOM00;
-  while (p->customs[type - TCOD_TYPE_CUSTOM00] && type < TCOD_TYPE_CUSTOM15) type = (TCOD_value_type_t)(type + 1);
-  if (p->customs[type - TCOD_TYPE_CUSTOM00]) {
+  while (parser->customs[type - TCOD_TYPE_CUSTOM00] && type < TCOD_TYPE_CUSTOM15) type = (TCOD_value_type_t)(type + 1);
+  if (parser->customs[type - TCOD_TYPE_CUSTOM00]) {
     /* no more custom types slots available */
     return TCOD_TYPE_NONE;
   }
-  p->customs[type - TCOD_TYPE_CUSTOM00] = custom_type_parser;
+  parser->customs[type - TCOD_TYPE_CUSTOM00] = custom_type_parser;
   return type;
 }
 
-TCOD_parser_struct_t TCOD_parser_new_struct(TCOD_parser_t parser, const char* name) {
-  TCOD_struct_int_t* ent = (TCOD_struct_int_t*)calloc(1, sizeof(TCOD_struct_int_t));
+TCOD_ParserStruct* TCOD_parser_new_struct(TCOD_Parser* parser, const char* name) {
+  TCOD_ParserStruct* ent = calloc(1, sizeof(*ent));
   ent->name = TCOD_strdup(name);
   ent->flags = TCOD_list_new();
   ent->props = TCOD_list_new();
   ent->lists = TCOD_list_new();
   ent->structs = TCOD_list_new();
-  TCOD_list_push(((TCOD_parser_int_t*)parser)->structs, ent);
-  return (TCOD_parser_struct_t)ent;
+  TCOD_list_push(parser->structs, ent);
+  return ent;
 }
 
-void TCOD_parser_delete(TCOD_parser_t parser) {
-  TCOD_parser_int_t* p = (TCOD_parser_int_t*)parser;
-  TCOD_struct_int_t** i_def;
-  TCOD_struct_prop_t** propCleanup;
-
-  char*** listCleanup;
-  int listSize = 0;
-
-  for (i_def = (TCOD_struct_int_t**)TCOD_list_begin(p->structs);
-       i_def != (TCOD_struct_int_t**)TCOD_list_end(p->structs);
-       i_def++) {
+void TCOD_parser_delete(TCOD_Parser* parser) {
+  for (TCOD_ParserStruct** i_def = (void*)TCOD_list_begin(parser->structs);
+       i_def != (void*)TCOD_list_end(parser->structs);
+       ++i_def) {
     free((*i_def)->name);
 
-    for (propCleanup = (TCOD_struct_prop_t**)TCOD_list_begin((*i_def)->props);
-         propCleanup != (TCOD_struct_prop_t**)TCOD_list_end((*i_def)->props);
-         propCleanup++) {
+    for (TCOD_struct_prop_t** propCleanup = (void*)TCOD_list_begin((*i_def)->props);
+         propCleanup != (void*)TCOD_list_end((*i_def)->props);
+         ++propCleanup) {
       free((*propCleanup)->name);
     }
     TCOD_list_clear_and_delete((*i_def)->props);
 
-    for (listCleanup = (char***)TCOD_list_begin((*i_def)->lists);
-         listCleanup != (char***)TCOD_list_end((*i_def)->lists);
-         listCleanup++) {
+    int listSize = 0;
+    for (char*** listCleanup = (void*)TCOD_list_begin((*i_def)->lists);
+         listCleanup != (void*)TCOD_list_end((*i_def)->lists);
+         ++listCleanup) {
       while ((*listCleanup)[listSize] != NULL) {
         free((*listCleanup)[listSize]);
-        listSize++;
+        ++listSize;
       }
     }
     TCOD_list_clear_and_delete((*i_def)->lists);
   }
-  TCOD_list_clear_and_delete(p->structs);
+  TCOD_list_clear_and_delete(parser->structs);
 }
 
 /* parse a file */
 static struct TCOD_List* default_props;
 /* triggers callbacks in the listener for each event during parsing */
-void TCOD_parser_run(TCOD_parser_t parser, const char* filename, TCOD_parser_listener_t* _listener) {
-  TCOD_parser_int_t* p = (TCOD_parser_int_t*)parser;
-  if (!_listener && !p->props) p->props = TCOD_list_new();
+void TCOD_parser_run(TCOD_Parser* parser, const char* filename, TCOD_parser_listener_t* _listener) {
+  if (!_listener && !parser->props) parser->props = TCOD_list_new();
   listener = _listener ? _listener : &default_listener;
-  default_props = p->props;
+  default_props = parser->props;
   lex = TCOD_lex_new(symbols, keywords, "//", "/*", "*/", NULL, "\"", TCOD_LEX_FLAG_NESTING_COMMENT);
   if (!TCOD_lex_set_data_file(lex, (char*)filename)) {
     char buf[1024];
@@ -740,15 +683,9 @@ void TCOD_parser_run(TCOD_parser_t parser, const char* filename, TCOD_parser_lis
     return;
   }
   while (1) {
-    bool named = false;
-    char id[BIG_NAME_LEN * 2 + 2];
-    char type[BIG_NAME_LEN];
-    TCOD_lex_t save;
-    TCOD_struct_int_t* def = NULL;
-    TCOD_struct_int_t** i_def;
-    bool dynStruct = false;
     TCOD_lex_parse(lex);
     if (lex->token_type == TCOD_LEX_EOF || lex->token_type == TCOD_LEX_ERROR) break;
+    bool dynStruct = false;
     if (lex->token_type == TCOD_LEX_KEYWORD) {
       if (strcmp(lex->tok, "struct") == 0) {
         /* level 0 dynamic structure declaration */
@@ -764,24 +701,29 @@ void TCOD_parser_run(TCOD_parser_t parser, const char* filename, TCOD_parser_lis
       TCOD_parser_error("Parser::parse : identifier token expected");
       return;
     }
-    string_copy(type, lex->tok, BIG_NAME_LEN);
+    char type[MAX_NAME_LEN];
+    string_copy(type, lex->tok, MAX_NAME_LEN);
+    char id[MAX_NAME_LEN * 2 + 2];
     strcpy(id, type);
+    TCOD_lex_t save;
     TCOD_lex_savepoint(lex, &save);
+    bool named = false;
     if (TCOD_lex_parse(lex) == TCOD_LEX_STRING) {
       /* named entity. id = <type>#<name> */
       strcat(id, "#");
-      if (strlen(lex->tok) >= BIG_NAME_LEN) {
-        TCOD_parser_error("Parser::parse : name %s too long. Max %d characters", lex->tok, BIG_NAME_LEN - 1);
+      if (strlen(lex->tok) >= MAX_NAME_LEN) {
+        TCOD_parser_error("Parser::parse : name %s too long. Max %d characters", lex->tok, MAX_NAME_LEN - 1);
         return;
       }
       strcat(id, lex->tok);
       named = true;
     }
     TCOD_lex_restore(lex, &save);
+    TCOD_ParserStruct* def = NULL;
     do {
       /* look for a definition for id */
-      for (i_def = (TCOD_struct_int_t**)TCOD_list_begin(p->structs);
-           i_def != (TCOD_struct_int_t**)TCOD_list_end(p->structs);
+      for (TCOD_ParserStruct** i_def = (void*)TCOD_list_begin(parser->structs);
+           i_def != (void*)TCOD_list_end(parser->structs);
            i_def++) {
         if (strcmp((*i_def)->name, id) == 0) {
           def = *i_def;
@@ -790,8 +732,8 @@ void TCOD_parser_run(TCOD_parser_t parser, const char* filename, TCOD_parser_lis
       }
       if (def == NULL && named) {
         /* look for general definition <type> for entity <type>#<name> */
-        for (i_def = (TCOD_struct_int_t**)TCOD_list_begin(p->structs);
-             i_def != (TCOD_struct_int_t**)TCOD_list_end(p->structs);
+        for (TCOD_ParserStruct** i_def = (void*)TCOD_list_begin(parser->structs);
+             i_def != (void*)TCOD_list_end(parser->structs);
              i_def++) {
           if (strcmp((*i_def)->name, type) == 0) {
             def = (*i_def);
@@ -808,8 +750,8 @@ void TCOD_parser_run(TCOD_parser_t parser, const char* filename, TCOD_parser_lis
       TCOD_parser_error("Parser::parse : unknown entity type %s", type);
       return;
     } else {
-      if (!listener->new_struct((TCOD_parser_struct_t)def, named ? strchr(id, '#') + 1 : NULL)) return;
-      if (!TCOD_parser_parse_entity(p, def)) return;
+      if (!listener->new_struct(def, named ? strchr(id, '#') + 1 : NULL)) return;
+      if (!TCOD_parser_parse_entity(parser, def)) return;
     }
   }
   if (lex->token_type == TCOD_LEX_ERROR) {
@@ -825,17 +767,17 @@ typedef struct {
   TCOD_value_t value;
 } prop_t;
 static char cur_prop_name[512] = "";
-static bool default_new_struct(TCOD_parser_struct_t str, const char* name) {
+static bool default_new_struct(TCOD_ParserStruct* str, const char* name) {
   (void)name;  // Ignored parameters.
   if (cur_prop_name[0]) strcat(cur_prop_name, ".");
-  strcat(cur_prop_name, ((TCOD_struct_int_t*)str)->name);
+  strcat(cur_prop_name, str->name);
   return true;
 }
 
 static bool default_new_flag(const char* name) {
   char tmp[512] = "";
-  prop_t* prop = (prop_t*)calloc(sizeof(prop_t), 1);
   snprintf(tmp, sizeof(tmp), "%s.%s", cur_prop_name, name);
+  prop_t* prop = calloc(sizeof(*prop), 1);
   prop->name = TCOD_strdup(tmp);
   prop->type = TCOD_TYPE_BOOL;
   prop->value.b = true;
@@ -845,8 +787,8 @@ static bool default_new_flag(const char* name) {
 
 static bool default_new_property(const char* propname, TCOD_value_type_t type, TCOD_value_t value) {
   char tmp[512] = "";
-  prop_t* prop = (prop_t*)calloc(sizeof(prop_t), 1);
   snprintf(tmp, sizeof(tmp), "%s.%s", cur_prop_name, propname);
+  prop_t* prop = calloc(sizeof(*prop), 1);
   prop->name = TCOD_strdup(tmp);
   prop->type = type;
   prop->value = value;
@@ -854,7 +796,7 @@ static bool default_new_property(const char* propname, TCOD_value_type_t type, T
   return true;
 }
 
-static bool default_end_struct(TCOD_parser_struct_t str, const char* name) {
+static bool default_end_struct(TCOD_ParserStruct* str, const char* name) {
   (void)str;
   (void)name;  // Ignored parameters.
   char* ptr = strrchr(cur_prop_name, '.');
@@ -867,16 +809,11 @@ static bool default_end_struct(TCOD_parser_struct_t str, const char* name) {
 
 static void default_error(const char* msg) { TCOD_set_errorvf("Error during parser: %s\n", msg); }
 
-static const TCOD_value_t* TCOD_get_property(TCOD_parser_t parser, TCOD_value_type_t expectedType, const char* name) {
-  void** it;
-  char tmp[512], err[512];
-  char *ptr, *curname;
-  TCOD_struct_int_t* str = NULL;
-  TCOD_value_type_t type;
-  TCOD_parser_int_t* p = (TCOD_parser_int_t*)parser;
-  if (!p->props) return NULL;
-  for (it = TCOD_list_begin(p->props); it != TCOD_list_end(p->props); it++) {
-    prop_t* prop = *((prop_t**)it);
+static const TCOD_value_t* TCOD_get_property(TCOD_Parser* parser, TCOD_value_type_t expectedType, const char* name) {
+  if (!parser->props) return NULL;
+  char err[512];
+  for (prop_t** it = (void*)TCOD_list_begin(parser->props); it != (void*)TCOD_list_end(parser->props); ++it) {
+    prop_t* prop = *it;
     if (strcmp(prop->name, name) == 0) {
       /* property found. check type */
       if (expectedType == TCOD_TYPE_STRING && prop->type >= TCOD_TYPE_VALUELIST00 &&
@@ -892,29 +829,33 @@ static const TCOD_value_t* TCOD_get_property(TCOD_parser_t parser, TCOD_value_ty
     }
   }
   /* property not found. Check if it exists */
+  char tmp[512];
   strcpy(tmp, name);
-  ptr = strchr(tmp, '.');
-  curname = tmp;
   sprintf(err, "Fatal error ! Try to read unknown property '%s'\n", name);
+  char* ptr = strchr(tmp, '.');
+  char* current_name = tmp;
+  TCOD_ParserStruct* str = NULL;
   while (ptr) {
     bool found = false;
     *ptr = 0;
-    for (it = TCOD_list_begin(p->structs); !found && it != TCOD_list_end(p->structs); it++) {
-      str = *((TCOD_struct_int_t**)it);
-      if (strcmp(str->name, curname) == 0) found = true;
+    for (TCOD_ParserStruct** it = (void*)TCOD_list_begin(parser->structs);
+         !found && it != (void*)TCOD_list_end(parser->structs);
+         ++it) {
+      str = *it;
+      if (strcmp(str->name, current_name) == 0) found = true;
     }
     if (!found) {
       /* one of the structures is unknown */
       default_error(err);
     }
-    curname = ptr + 1;
-    ptr = strchr(curname, '.');
+    current_name = ptr + 1;
+    ptr = strchr(current_name, '.');
   }
   if (!str) {
     /* no structure in name */
     default_error(err);
   }
-  type = TCOD_struct_get_type((TCOD_parser_struct_t)str, curname);
+  TCOD_value_type_t type = TCOD_struct_get_type(str, current_name);
   if (type == TCOD_TYPE_NONE) {
     /* property does not exist in structure */
     default_error(err);
@@ -923,68 +864,63 @@ static const TCOD_value_t* TCOD_get_property(TCOD_parser_t parser, TCOD_value_ty
   return NULL;
 }
 
-bool TCOD_parser_has_property(TCOD_parser_t parser, const char* name) {
-  void** it;
-  TCOD_parser_int_t* p = (TCOD_parser_int_t*)parser;
-  if (!p->props) return false;
-  for (it = TCOD_list_begin(p->props); it != TCOD_list_end(p->props); it++) {
-    prop_t* prop = *((prop_t**)it);
-    if (strcmp(prop->name, name) == 0) {
+bool TCOD_parser_has_property(TCOD_Parser* parser, const char* name) {
+  if (!parser->props) return false;
+  for (prop_t** prop = (void*)TCOD_list_begin(parser->props); prop != (void*)TCOD_list_end(parser->props); ++prop) {
+    if (strcmp((*prop)->name, name) == 0) {
       return true;
     }
   }
   return false;
 }
 
-bool TCOD_parser_get_bool_property(TCOD_parser_t parser, const char* name) {
+bool TCOD_parser_get_bool_property(TCOD_Parser* parser, const char* name) {
   const TCOD_value_t* value = TCOD_get_property(parser, TCOD_TYPE_BOOL, name);
   return value ? value->b : false;
 }
 
-int TCOD_parser_get_int_property(TCOD_parser_t parser, const char* name) {
+int TCOD_parser_get_int_property(TCOD_Parser* parser, const char* name) {
   const TCOD_value_t* value = TCOD_get_property(parser, TCOD_TYPE_INT, name);
   return value ? value->i : 0;
 }
 
-int TCOD_parser_get_char_property(TCOD_parser_t parser, const char* name) {
+int TCOD_parser_get_char_property(TCOD_Parser* parser, const char* name) {
   const TCOD_value_t* value = TCOD_get_property(parser, TCOD_TYPE_CHAR, name);
   return value ? value->c : 0;
 }
 
-float TCOD_parser_get_float_property(TCOD_parser_t parser, const char* name) {
+float TCOD_parser_get_float_property(TCOD_Parser* parser, const char* name) {
   const TCOD_value_t* value = TCOD_get_property(parser, TCOD_TYPE_FLOAT, name);
   return value ? value->f : 0.0f;
 }
 
-const char* TCOD_parser_get_string_property(TCOD_parser_t parser, const char* name) {
+const char* TCOD_parser_get_string_property(TCOD_Parser* parser, const char* name) {
   const TCOD_value_t* value = TCOD_get_property(parser, TCOD_TYPE_STRING, name);
   return value ? value->s : NULL;
 }
 
-TCOD_color_t TCOD_parser_get_color_property(TCOD_parser_t parser, const char* name) {
+TCOD_color_t TCOD_parser_get_color_property(TCOD_Parser* parser, const char* name) {
   const TCOD_value_t* value = TCOD_get_property(parser, TCOD_TYPE_COLOR, name);
   return value ? value->col : TCOD_black;
 }
 
-TCOD_dice_t TCOD_parser_get_dice_property(TCOD_parser_t parser, const char* name) {
-  static TCOD_dice_t default_dice = {0, 0, 0.0f, 0.0f};
+TCOD_dice_t TCOD_parser_get_dice_property(TCOD_Parser* parser, const char* name) {
   const TCOD_value_t* value = TCOD_get_property(parser, TCOD_TYPE_DICE, name);
-  return value ? value->dice : default_dice;
+  return value ? value->dice : (TCOD_dice_t){0, 0, 0.0f, 0.0f};
 }
 
-void TCOD_parser_get_dice_property_py(TCOD_parser_t parser, const char* name, TCOD_dice_t* dice) {
+void TCOD_parser_get_dice_property_py(TCOD_Parser* parser, const char* name, TCOD_dice_t* dice) {
   *dice = TCOD_parser_get_dice_property(parser, name);
 }
 
-void* TCOD_parser_get_custom_property(TCOD_parser_t parser, const char* name) {
+void* TCOD_parser_get_custom_property(TCOD_Parser* parser, const char* name) {
   const TCOD_value_t* value = TCOD_get_property(parser, TCOD_TYPE_CUSTOM00, name);
   return value ? value->custom : NULL;
 }
 
-TCOD_list_t TCOD_parser_get_list_property(TCOD_parser_t parser, const char* name, TCOD_value_type_t type) {
+TCOD_list_t TCOD_parser_get_list_property(TCOD_Parser* parser, const char* name, TCOD_value_type_t type) {
   static TCOD_list_t empty_list = NULL;
-  const TCOD_value_t* value;
   if (!empty_list) empty_list = TCOD_list_new();
-  value = TCOD_get_property(parser, TCOD_TYPE_LIST | type, name);
+  const TCOD_value_t* value = TCOD_get_property(parser, TCOD_TYPE_LIST | type, name);
   return value ? value->list : empty_list;
 }